requirements.txt
aiofiles==*
aiogram==*
aiohappyeyeballs==*
aiohttp==*
aiosignal==*
annotated-types==*
anyio==*
attrs==*
certifi==*
frozenlist==*
h11==*
httpcore==*
httpx==*
idna==*
magic-filter==*
multidict==*
propcache==*
pydantic==*
pydantic_core==*
python-dotenv==*
ruff==*
setuptools==*
sniffio==*
typing-inspection==*
typing_extensions==*
uvloop==*
wheel==*
yarl==*
